Change responsive height to HoC
