+++ src/index.js
@@ -1,1 +1,1 @@
-export { default } from "./components/multi_select_height_responsive";
+export { default } from "./components/multi_select";
